+++ CHANGELOG.md
@@ -38,6 +38,7 @@
   + Warns when right-clicking with a Wither Impact weapon (e.g. Hyperion) no longer gains combat exp
   (Kill a mob with melee-hits to fix this hypixel bug)
 + Added Piece of Wizard Portal show earned by player name (Contributed by HiZe)
++ City Project Daily Reminder - Remind every 24 hours to participate
 
 ### Changes
 + Added Options for displays Crop Milestone and Best Crop Time.